popovers: Move scroll handler to popover_menus module.
Part of cleaning out the original legacy popovers system.
