+++ web/src/click_handlers.js
@@ -669,7 +669,6 @@ export function initialize() {
         });
     }
 
-    popovers.register_click_handlers();
     user_profile.register_click_handlers();
     stream_popover.register_click_handlers();
